+++ dalesbred-junit/src/test/java/fi/evident/dalesbred/junit/TransactionalTestsTest.java
@@ -34,7 +34,7 @@ import static org.junit.Assert.assertEquals;
 
 public class TransactionalTestsTest {
 
-    private final Database db = TestDatabaseProvider.databaseForProperties("connection.properties");
+    private final Database db = Database.forUrlAndCredentials("jdbc:hsqldb:.", "sa", "");
 
     @Rule
     public final TransactionalTests transactionalTests = new TransactionalTests(db);