+++ dalesbred/src/test/java/fi/evident/dalesbred/DatabaseDataSourceTest.java
@@ -31,8 +31,8 @@ public class DatabaseDataSourceTest {
 
     @Test
     public void creatingDatabaseFromDataSource() {
-        Database db = Database.forDataSource(TestDatabaseProvider.createDataSource());
+        Database db = Database.forDataSource(TestDatabaseProvider.createInMemoryHSQLDataSource());
 
-        assertThat(db.findUniqueInt("select 42"), is(42));
+        assertThat(db.findUniqueInt("values (42)"), is(42));
     }
 }