+++ dalesbred/src/test/java/fi/evident/dalesbred/DatabaseResultTableTest.java
@@ -36,30 +36,30 @@ import static org.junit.Assert.assertThat;
 
 public class DatabaseResultTableTest {
 
-    private final Database db = TestDatabaseProvider.createTestDatabase();
+    private final Database db = TestDatabaseProvider.createInMemoryHSQLDatabase();
 
     @Rule
     public final TransactionalTestsRule rule = new TransactionalTestsRule(db);
 
     @Test
     public void fetchSimpleResultTable() {
-        ResultTable table = db.findTable("select 42 as num, 'foo' as str, true as bool");
+        ResultTable table = db.findTable("select 42 as num, 'foo' as str, true as bool from (values (0)) v");
 
         assertThat(table.getColumnCount(), is(3));
-        assertThat(table.getColumnNames(), is(asList("num", "str", "bool")));
+        assertThat(table.getColumnNames(), is(asList("NUM", "STR", "BOOL")));
         assertThat(table.getColumnTypes(), is(types(Integer.class, String.class, Boolean.class)));
-        assertThat(table.getColumns().toString(), is("[num: java.lang.Integer, str: java.lang.String, bool: java.lang.Boolean]"));
+        assertThat(table.getColumns().toString(), is("[NUM: java.lang.Integer, STR: java.lang.String, BOOL: java.lang.Boolean]"));
         assertThat(table.getColumns().get(1).getIndex(), is(1));
 
         assertThat("jdbcType[0]", table.getColumns().get(0).getJdbcType(), is(Types.INTEGER));
-        assertThat("databaseType[0]", table.getColumns().get(0).getDatabaseType(), is("int4"));
+        assertThat("databaseType[0]", table.getColumns().get(0).getDatabaseType(), is("INTEGER"));
 
         assertThat(table.getRowCount(), is(1));
         assertThat(table.getRows().get(0).asList(), is(values(42, "foo", true)));
         assertEquals("foo", table.get(0, 1));
         assertEquals("foo", table.get(0, "str"));
 
-        assertThat(table.toString(), is("ResultTable [columns=[num: java.lang.Integer, str: java.lang.String, bool: java.lang.Boolean], rows=1]"));
+        assertThat(table.toString(), is("ResultTable [columns=[NUM: java.lang.Integer, STR: java.lang.String, BOOL: java.lang.Boolean], rows=1]"));
     }
 
     @NotNull