+++ dalesbred/src/test/java/fi/evident/dalesbred/support/guice/DriverManagerDatabaseModuleTest.java
@@ -36,10 +36,10 @@ public class DriverManagerDatabaseModuleTest {
     @Test
     public void testCreation() {
         TestDatabaseProvider.assumeConfigurationExists();
-        Injector injector = Guice.createInjector(new DriverManagerDatabaseModule(), TestDatabaseProvider.propertiesModule());
+        Injector injector = Guice.createInjector(new DriverManagerDatabaseModule(), TestDatabaseProvider.inMemoryDatabasePropertiesModule());
 
         Database db = injector.getInstance(Database.class);
 
-        assertThat(db.findUniqueInt("select 42"), is(42));
+        assertThat(db.findUniqueInt("values (42)"), is(42));
     }
 }